Moved swipeguestures to the calendar control; Added patter for proper even subscription/unsubscription Cleanup
