+++ samples/SampleApp/Views/MainPage.xaml
@@ -23,7 +23,7 @@
                 Light="#000000"/>
 
             <Style x:Key="LabelMainPageStyle"
-                   TargetType="Label">
+                    TargetType="Label">
                 <Setter Property="FontAttributes"
                         Value="Bold"/>
                 <Setter Property="FontSize"
@@ -36,7 +36,7 @@
                         Value="{toolkit:AppThemeResource LabelColor}"/>
             </Style>
             <Style x:Key="ButtonMainPageStyle"
-                   TargetType="Button">
+                    TargetType="Button">
                 <Setter Property="HorizontalOptions"
                         Value="Fill"/>
                 <Setter Property="BackgroundColor"
@@ -62,15 +62,9 @@
     </ContentPage.Resources>
     <ScrollView>
         <VerticalStackLayout HorizontalOptions="Fill"
-                             Spacing="6">
-            <Button
-                Margin="30,30,30,10"
-                Clicked="NeoCalendar"
-                Style="{StaticResource ButtonMainPageStyle}"
-                Text="Neo Calendar"/>
-            <Label
-                Style="{StaticResource LabelMainPageStyle}"
-                Text="Event Calendar"/>
+                Spacing="6">
+            <Label Style="{StaticResource LabelMainPageStyle}"
+                    Text="Event Calendar"/>
             <Button
                 Clicked="SimpleCalendar"
                 Style="{StaticResource ButtonMainPageStyle}"
@@ -92,13 +86,13 @@
                 Style="{StaticResource ButtonMainPageStyle}"
                 Text="Range Event Calendar"/>
             <Label Style="{StaticResource LabelMainPageStyle}"
-                   Text="Calendars from Devices"/>
+                    Text="Calendars from Devices"/>
             <Button
                 Clicked="Windows11Calendar"
                 Style="{StaticResource ButtonMainPageStyle}"
                 Text="Windows 11 Calendar"/>
             <Label Style="{StaticResource LabelMainPageStyle}"
-                   Text="Picker Popup"/>
+                    Text="Picker Popup"/>
             <Button
                 Clicked="PickerPopup"
                 Style="{StaticResource ButtonMainPageStyle}"
@@ -112,7 +106,7 @@
                 Style="{StaticResource ButtonMainPageStyle}"
                 Text="Range Start-End Selection Calendar"/>
             <Label Style="{StaticResource LabelMainPageStyle}"
-                   Text="Week View Calendars"/>
+                    Text="Week View Calendars"/>
             <Button
                 Clicked="WeekViewCalendar"
                 Style="{StaticResource ButtonMainPageStyle}"
