--- samples/SampleApp/Views/NeoCalendarPage.xaml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage
-        x:Class="SampleApp.Views.NeoCalendarPage"
-        xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-        xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-        xmlns:local="clr-namespace:SampleApp.ViewModels"
-        xmlns:plugin="clr-namespace:Plugin.Maui.Calendar.Controls;assembly=Plugin.Maui.Calendar"
-        xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
-        x:Name="simpleCalendarPage"
-        x:DataType="local:NeoCalendarPageViewModel"
-        Shell.TabBarIsVisible="False"
-        BackgroundColor="{toolkit:AppThemeResource BackgroundColor}">
-        <ContentPage.Resources>
-                <ResourceDictionary>
-                        <Style x:Key="daysLabelStyle"
-                               TargetType="Label">
-                                <Setter Property="HorizontalTextAlignment"
-                                        Value="Center"/>
-                                <Setter Property="TextColor"
-                                        Value="Black"/>
-                                <Setter Property="FontSize"
-                                        Value="Medium"/>
-                                <Setter Property="LineBreakMode"
-                                        Value="WordWrap"/>
-                                <Setter Property="VerticalTextAlignment"
-                                        Value="Center"/>
-                                <Setter Property="Margin"
-                                        Value="5,2"/>
-                        </Style>
-                        <Style x:Key="daysTitleLabelStyle"
-                               TargetType="Label">
-                                <Setter Property="FontSize"
-                                        Value="18"/>
-                        </Style>
-                </ResourceDictionary>
-        </ContentPage.Resources>
-        <ContentPage.ToolbarItems>
-                <ToolbarItem Command="{Binding TodayCommand}"
-                             Text="Today"/>
-        </ContentPage.ToolbarItems>
-        <plugin:CalendarNeo
-                x:Name="calendar"
-                Padding="10,0"
-                ArrowsFontSize="Medium"
-                DaysTitleLabelStyle="{StaticResource daysTitleLabelStyle}"
-                HorizontalOptions="Fill"
-                Month="Jan"
-                MonthIndex="{Binding Month}"
-                VerticalOptions="Fill"
-                Year="{Binding Year}"/>
-</ContentPage>
