--- samples/SampleApp/Views/NeoCalendarPage.xaml.cs
@@ -1,9 +1,0 @@
-namespace SampleApp.Views;
-
-public partial class NeoCalendarPage : ContentPage
-{
-	public NeoCalendarPage()
-	{
-		InitializeComponent();
-	}
-}
