+++ src/Plugin.Maui.Calendar/Shared/Controls/MonthDaysView.cs
@@ -13,24 +13,6 @@ namespace Plugin.Maui.Calendar.Controls;
 
 public partial class MonthDaysView : ContentView
 {
-    SwipeGestureRecognizer leftSwipeGesture;
-    SwipeGestureRecognizer rightSwipeGesture;
-    SwipeGestureRecognizer upSwipeGesture;
-    SwipeGestureRecognizer downSwipeGesture;
-
-    public static readonly BindableProperty SwipeDetectionDisabledProperty =
-        BindableProperty.Create(
-            nameof(SwipeDetectionDisabled),
-            typeof(bool),
-            typeof(MonthDaysView),
-            false
-        );
-
-    public bool SwipeDetectionDisabled
-    {
-        get => (bool)GetValue(SwipeDetectionDisabledProperty);
-        set => SetValue(SwipeDetectionDisabledProperty, value);
-    }
 
     private Grid _daysControl;
 
@@ -840,50 +822,9 @@ public partial class MonthDaysView : ContentView
 
     public MonthDaysView()
     {
-        if (!SwipeDetectionDisabled)
-        {
-            leftSwipeGesture = new() { Direction = SwipeDirection.Left };
-            rightSwipeGesture = new() { Direction = SwipeDirection.Right };
-            upSwipeGesture = new() { Direction = SwipeDirection.Up };
-            downSwipeGesture = new() { Direction = SwipeDirection.Down };
-
-            Loaded += LoadedMethod;
-            Unloaded += UnloadedMethod;
-        }
         RenderLayout();
     }
 
-    private void LoadedMethod(object sender, EventArgs e)
-    {
-        if (!SwipeDetectionDisabled)
-        {
-            GestureRecognizers.Add(leftSwipeGesture);
-            GestureRecognizers.Add(rightSwipeGesture);
-            GestureRecognizers.Add(upSwipeGesture);
-            GestureRecognizers.Add(downSwipeGesture);
-
-            leftSwipeGesture.Swiped += OnSwiped;
-            rightSwipeGesture.Swiped += OnSwiped;
-            upSwipeGesture.Swiped += OnSwiped;
-            downSwipeGesture.Swiped += OnSwiped;
-        }
-    }
-
-    private void UnloadedMethod(object sender, EventArgs e)
-    {
-        if (!SwipeDetectionDisabled && GestureRecognizers.Count > 0)
-        {
-            GestureRecognizers.Remove(leftSwipeGesture);
-            GestureRecognizers.Remove(rightSwipeGesture);
-            GestureRecognizers.Remove(upSwipeGesture);
-            GestureRecognizers.Remove(downSwipeGesture);
-
-            leftSwipeGesture.Swiped -= OnSwiped;
-            rightSwipeGesture.Swiped -= OnSwiped;
-            upSwipeGesture.Swiped -= OnSwiped;
-            downSwipeGesture.Swiped -= OnSwiped;
-        }
-    }
 
     #region PropertyChanged
 
@@ -1115,10 +1056,6 @@ public partial class MonthDaysView : ContentView
 
     private void DiposeDayViews()
     {
-        leftSwipeGesture = null;
-        rightSwipeGesture = null;
-        upSwipeGesture = null;
-        downSwipeGesture = null;
 
         foreach (var dayView in _daysControl.Children.OfType<DayView>())
         {
@@ -1215,42 +1152,10 @@ public partial class MonthDaysView : ContentView
         }
     }
 
-    void OnSwiped(object sender, SwipedEventArgs e)
-    {
-        switch (e.Direction)
-        {
-            case SwipeDirection.Left:
-                OnSwipeLeft();
-                break;
-            case SwipeDirection.Right:
-                OnSwipeRight();
-                break;
-            case SwipeDirection.Up:
-                OnSwipeUp();
-                break;
-            case SwipeDirection.Down:
-                OnSwipeDown();
-                break;
-        }
-    }
 
-    public event EventHandler SwipedLeft;
-    public event EventHandler SwipedRight;
-    public event EventHandler SwipedUp;
-    public event EventHandler SwipedDown;
-
-    void OnSwipeLeft() => SwipedLeft?.Invoke(this, EventArgs.Empty);
-
-    void OnSwipeRight() => SwipedRight?.Invoke(this, EventArgs.Empty);
-
-    void OnSwipeUp() => SwipedUp?.Invoke(this, EventArgs.Empty);
-
-    void OnSwipeDown() => SwipedDown?.Invoke(this, EventArgs.Empty);
 
     public void Dispose()
     {
-        Loaded -= LoadedMethod;
-        Unloaded -= UnloadedMethod;
         DiposeDayViews();
     }
 }
